Support for Profiles (Saved cards) Feature
This feature required some hefty additions including new classes for both the
server and client as well as some revised handling of the `$request` array
in order to handle the various potential transaction types:

1. Regular transaction
2. User saves card for the first time
3. User saves an additional card
4. User uses a saved card

The primary complicating factor is that a new profile can be created while
also capturing a transaction, easy enough. The process is different, however
for adding a card to an existing profile-- this requires an API call ahead of
the Auth to add the card to the profile.